backend/requirements.txt
fastapi
uvicorn
asyncpg 
python-socketio